+++ src/irmin-pack/unix/irmin_pack_unix.ml
@@ -1,4 +1,5 @@
 module Dict = Pack_dict
+module Stats = Stats
 module Index = Pack_index
 module Inode = Inode
 module Pack_store = Pack_store