+++ CHANGES.md
@@ -1,3 +1,11 @@
+## Unreleased
+
+### Added
+
+- **irmin**
+  - Add `Metrics` module to describe metric gathering in irmin.
+    (#1817, @maiste)
+
 ## 3.2.1 (2022-04-07)
 
 - Support all version of cmdliner (#1803, @samoht)